Refactor Chart
Cleaned up chart component by seperating component into sepererate function, component, data files ect...
